GenericBoosting classifier

GenericBoosting classifier

GenericBoostingClassifier(
  base_model = NULL,
  n_estimators = 100L,
  learning_rate = 0.1,
  n_hidden_features = 5L,
  reg_lambda = 0.1,
  row_sample = 1,
  col_sample = 1,
  dropout = 0,
  tolerance = 1e-04,
  direct_link = 1L,
  verbose = 1L,
  seed = 123L,
  activation = "relu",
  n_clusters = 0,
  clustering_method = "kmeans",
  cluster_scaling = "standard",
  degree = NULL,
  weights_distr = "uniform"
)

Arguments

base_model:

object, base model to be boosted.

n_estimators:

int, number of boosting iterations.

learning_rate:

float, controls the learning speed at training time.

n_hidden_features:

int

number

of nodes in successive hidden layers.

reg_lambda:

float, L2 regularization parameter for successive errors in the optimizer (at training time).

row_sample:

float, percentage of rows chosen from the training set.

col_sample:

float, percentage of columns chosen from the training set.

dropout:

float, percentage of nodes dropped from the training set.

tolerance:

float, controls early stopping in gradient descent (at training time).

direct_link:

bool, indicates whether the original features are included (True) in model's fitting or not (False).

verbose:

int, progress bar (yes = 1) or not (no = 0) (currently).

seed:

int, reproducibility seed for nodes_sim=='uniform', clustering and dropout.

activation:

str, activation function: currently 'relu', 'relu6', 'sigmoid', 'tanh'

n_clusters:

int, number of clusters for clustering.

clustering_method:

str, clustering method: currently 'kmeans', 'gmm' (Gaussian Mixture Model)

cluster_scaling:

str, scaling method for clustering: currently 'standard', 'minmax', 'robust'

degree:

int, degree of polynomial interactions features.

weights_distr:

str, distribution of weights for the hidden layer currently 'uniform', 'gaussian'

Value

An object of class GenericBoostingClassifier

Examples

library(datasets)

X <- as.matrix(iris[, 1:4])
y <- as.integer(iris[, 5]) - 1L

n <- dim(X)[1]
p <- dim(X)[2]
set.seed(21341)
train_index <- sample(x = 1:n, size = floor(0.8*n), replace = TRUE)
test_index <- -train_index
X_train <- as.matrix(X[train_index, ])
y_train <- as.integer(y[train_index])
X_test <- as.matrix(X[test_index, ])
y_test <- as.integer(y[test_index])

obj <- mlsauce::GenericBoostingClassifier()

print(obj$get_params())
#> $activation
#> [1] "relu"
#> 
#> $backend
#> [1] "cpu"
#> 
#> $base_model
#> NULL
#> 
#> $bins
#> [1] "auto"
#> 
#> $cluster_scaling
#> [1] "standard"
#> 
#> $clustering_method
#> [1] "kmeans"
#> 
#> $col_sample
#> [1] 1
#> 
#> $degree
#> NULL
#> 
#> $direct_link
#> [1] 1
#> 
#> $dropout
#> [1] 0
#> 
#> $hist
#> [1] FALSE
#> 
#> $learning_rate
#> [1] 0.1
#> 
#> $n_clusters
#> [1] 0
#> 
#> $n_estimators
#> [1] 100
#> 
#> $n_hidden_features
#> [1] 5
#> 
#> $row_sample
#> [1] 1
#> 
#> $seed
#> [1] 123
#> 
#> $tolerance
#> [1] 1e-04
#> 
#> $verbose
#> [1] 1
#> 
#> $weights_distr
#> [1] "uniform"
#> 

obj$fit(X_train, y_train)
#> GenericBoostingClassifier(base_model=None, col_sample=1.0, dropout=0.0,
#>                           n_clusters=0.0, row_sample=1.0)

print(obj$score(X_test, y_test))
#> [1] 0.9253731